+++ VulkanEngine/Renderer/Resources/Mesh/Vertex.h
@@ -28,7 +28,7 @@ namespace Renderer
 		static std::vector<uint32_t> GetQuadIndecies();
 
 		static std::vector<Vertex> GetCubeVirticies();
-		static std::vector<uint16_t> GetCubeIndecies();
+		static std::vector<uint32_t> GetCubeIndecies();
 	};
